requirements/depends.txt
PyYAML==3.10
WebOb==1.2.3
amqp==1.0.13
anyjson==0.3.3
argh==0.23.3
beautifulsoup4==4.3.1
billiard==2.7.3.32
celery==3.0.22
django-nose==1.2
django-social-auth==0.7.25
httplib2==0.8
kombu==2.5.14
logilab-astng==0.24.3
logilab-common==0.60.0
meld3==0.6.10
oauth2==1.5.211
pathtools==0.1.2
python-dateutil==2.1
python-openid==2.2.5
sexpdata==0.0.3
six==1.3.0
supervisor==3.0
waitress==0.8.7
watchdog==0.6.0
wsgiref==0.1.2
